Move some of `coconut` helpers to `dock_crypto_utils`
