+++ coconut/Cargo.toml
@@ -17,11 +17,11 @@ ark-std = { version = "0.4.0", default-features = false }
 ark-poly = { version = "0.4.1", default-features = false }
 ark-serialize = { version = "0.4.1", default-features = false, features = [ "derive" ] }
 serde = { version = "1.0.156", default-features = false, features = ["derive"] }
-dock_crypto_utils = { version = "0.9.0", default-features = false }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
 digest = "0.10.6"
 zeroize = { version = "1.5.7", features = ["derive"] }
 rayon = { version = "1.7.0", optional = true }
-schnorr_pok = { version = "0.9.0", default-features = false }
+schnorr_pok = { version = "0.9.0", default-features = false, path = "../schnorr_pok" }
 itertools = "0.10.5"
 secret_sharing_and_dkg = { version = "0.2.0", default-features = false }
 serde_with = { version = "1.10.0", default-features = false, features = ["macros"] }
