+++ coconut/src/proof/messages_pok/mod.rs
@@ -7,6 +7,7 @@ use ark_ec::pairing::Pairing;
 use ark_serialize::{CanonicalSerialize, Write};
 use ark_std::{cfg_iter, rand::RngCore};
 
+use dock_crypto_utils::join;
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
 use schnorr_pok::{error::SchnorrError, SchnorrChallengeContributor};
@@ -14,7 +15,7 @@ use schnorr_pok::{error::SchnorrError, SchnorrChallengeContributor};
 use super::UnpackedBlindedMessages;
 use crate::{
     helpers::{schnorr_error, DoubleEndedExactSizeIterator, WithSchnorrAndBlindings},
-    join, pairs,
+    pairs,
     setup::SignatureParams,
     signature::message_commitment::MessageCommitmentRandomness,
     CommitMessage,
