+++ coconut/src/proof/signature_pok/proof.rs
@@ -1,6 +1,7 @@
 use alloc::vec::Vec;
 use ark_ec::pairing::Pairing;
 use ark_serialize::*;
+use dock_crypto_utils::join;
 use dock_crypto_utils::randomized_pairing_check::RandomizedPairingChecker;
 use serde::{Deserialize, Serialize};
