+++ coconut/src/setup/signature_params.rs
@@ -12,7 +12,7 @@ use serde_with::serde_as;
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 
-use crate::{concat_slices, join};
+use dock_crypto_utils::{concat_slices, join};
 
 /// Parameters generated by a random oracle.
 #[serde_as]
