+++ compressed_sigma/Cargo.toml
@@ -15,7 +15,7 @@ ark-std.workspace = true
 ark-poly.workspace = true
 rayon = {workspace = true, optional = true}
 digest.workspace = true
-dock_crypto_utils = { version = "0.9.0", default-features = false }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
 
 [dev-dependencies]
 blake2.workspace = true
