+++ legogroth16/Cargo.toml
@@ -29,7 +29,7 @@ num-bigint = { version = "0.4", default-features = false, optional = true }
 log = "0.4"
 ark-groth16 = { workspace = true, optional = true }
 ark-snark = { version = "^0.4.0", default-features = false, optional = true }
-dock_crypto_utils = { version = "0.9.0", default-features = false }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
 
 [dev-dependencies]
 csv = { version = "1" }
@@ -45,7 +45,6 @@ tracing = { version = "0.1", default-features = false, features = [ "attributes"
 tracing-subscriber = { version = "0.2" }
 hex-literal = "0.2.1"
 serde_json = "1.0.64"
-
 
 [profile.release]
 lto = true
