+++ proof_system/Cargo.toml
@@ -18,11 +18,11 @@ ark-ec.workspace = true
 ark-std.workspace = true
 digest.workspace = true
 rayon = {workspace = true, optional = true}
-bbs_plus = { version = "0.11.0", default-features = false }
-schnorr_pok = { version = "0.9.0", default-features = false }
-vb_accumulator = { version = "0.12.0", default-features = false }
-dock_crypto_utils = { version = "0.9.0", default-features = false }
-saver = { version = "0.9.0", default-features = false }
+bbs_plus = { version = "0.11.0", default-features = false, path = "../bbs_plus" }
+schnorr_pok = { version = "0.9.0", default-features = false, path = "../schnorr_pok" }
+vb_accumulator = { version = "0.12.0", default-features = false, path = "../vb_accumulator" }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
+saver = { version = "0.9.0", default-features = false, path = "../saver" }
 serde.workspace = true
 serde_with.workspace = true
 ark-groth16.workspace = true
@@ -31,7 +31,7 @@ ark-relations.workspace = true
 zeroize.workspace = true
 coconut-crypto = { version = "0.1.0", default-features = false, path = "../coconut" }
 merlin = { workspace = true, default-features = false }
-legogroth16 = { workspace = true, default-features = false, features = ["circom", "aggregation"] }
+legogroth16 = { version = "0.6.0", default-features = false, features = ["circom", "aggregation"], path = "../legogroth16" }
 itertools = "0.10.5"
 paste = "1.0.12"
