+++ saver/Cargo.toml
@@ -17,11 +17,11 @@ ark-relations.workspace = true
 ark-groth16.workspace = true
 digest.workspace = true
 rayon = {workspace = true, optional = true}
-dock_crypto_utils = { version = "0.9.0", default-features = false }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
 serde.workspace = true
 serde_with.workspace = true
 zeroize.workspace = true
-legogroth16 = { workspace = true, default-features = false, features = ["aggregation"] }
+legogroth16 = { version = "0.6.0", default-features = false, features = ["aggregation"], path = "../legogroth16" }
 merlin = { workspace = true, default-features = false }
 
 [dev-dependencies]
