+++ secret_sharing_and_dkg/Cargo.toml
@@ -18,8 +18,8 @@ rayon = {workspace = true, optional = true}
 serde.workspace = true
 serde_with.workspace = true
 zeroize.workspace = true
-dock_crypto_utils = { version = "0.9.0", default-features = false }
-schnorr_pok = { version = "0.9.0", default-features = false }
+dock_crypto_utils = { version = "0.9.0", default-features = false, path = "../utils" }
+schnorr_pok = { version = "0.9.0", default-features = false, path = "../schnorr_pok" }
 
 [dev-dependencies]
 blake2.workspace = true
