+++ test_utils/Cargo.toml
@@ -6,9 +6,9 @@ authors.workspace = true
 license.workspace = true
 
 [dependencies]
-bbs_plus = { version = "0.11.0", default-features = false }
-schnorr_pok = { version = "0.9.0", default-features = false }
-vb_accumulator = { version = "0.12.0", default-features = false }
+bbs_plus = { version = "0.11.0", default-features = false, path = "../bbs_plus" }
+schnorr_pok = { version = "0.9.0", default-features = false, path = "../schnorr_pok" }
+vb_accumulator = { version = "0.12.0", default-features = false, path = "../vb_accumulator" }
 ark-ff.workspace = true
 ark-ec.workspace = true
 ark-std.workspace = true
