+++ utils/src/extend_some.rs
