+++ utils/src/lib.rs
@@ -1,18 +1,16 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod aliases;
+pub mod extend_some;
 // TODO: Feature gate this
 #[macro_use]
 pub mod serde_utils;
 pub mod ff;
 pub mod hashing_utils;
+pub mod iter;
+pub mod macros;
 pub mod msm;
 pub mod poly;
 pub mod randomized_pairing_check;
 pub mod transcript;
-
-#[macro_export]
-macro_rules! concat_slices {
-    ($($slice: expr),+) => {
-        [$(&$slice[..]),+].concat()
-    }
-}
+pub mod try_iter;
