+++ utils/src/macros.rs
@@ -1,0 +1,209 @@
+/// Concatenates supplied slices into one continuous vector.
+#[macro_export]
+macro_rules! concat_slices {
+    ($($slice: expr),+) => {
+        [$(&$slice[..]),+].concat()
+    }
+}
+
+/// Implements `Deref`/`DeferMut` traits for the supplied wrapper and type.
+#[macro_export]
+macro_rules! impl_deref {
+    ($wrapper: ident$(<$($ty: ident: $($by: path),+),*>)?($type: path)) => {
+        impl$(<$($ty: $($by)++),+>)* core::ops::Deref for $wrapper$(<$($ty),+>)* {
+            type Target = $type;
+
+            fn deref(&self) -> &Self::Target {
+                &self.0
+            }
+        }
+
+        impl$(<$($ty: $($by)++),+>)* core::ops::DerefMut for $wrapper$(<$($ty),+>)* {
+            fn deref_mut(&mut self) -> &mut Self::Target {
+                &mut self.0
+            }
+        }
+    };
+}
+
+/// Calculates the product of pairing for supplied pairs.
+/// ```compile_fail
+/// multi_pairing! {
+///     a, c,
+///     b, d
+/// }
+/// ```
+/// Will be transformed to:
+/// ```compile_fail
+/// E::multi_pairing([a, b], [c, d])
+/// ```
+#[macro_export]
+macro_rules! multi_pairing {
+    ($($g1: expr, $g2: expr);+) => {
+        $crate::multi_pairing! { using E: $($g1, $g2);+ }
+    };
+    (using $pairing_engine: path: $($g1: expr, $g2: expr);+) => {
+        <$pairing_engine>::multi_pairing(
+            [
+                $($g1.into()),+
+            ],
+            [
+                $($g2.into()),+
+            ]
+        )
+    }
+}
+
+/// Flattened `rayon::join(|| expr1, || rayon::join(|| expr2, || ...))`
+#[cfg(feature = "parallel")]
+#[macro_export]
+macro_rules! join {
+    (@ $a: expr) => { $a };
+    (@ $a: expr, $b: expr) => {
+        rayon::join(|| $a, || $b)
+    };
+    (@ $a: expr, $b: expr, $($c: expr),+) => {{
+        join!(@ $a, join!(@ $b, $($c),+))
+    }};
+    ($($e: expr),+) => {{
+        $crate::unnest_tuple!(
+            $($e),+
+            =>
+            join!(@ $($e),+)
+        )
+    }}
+}
+
+/// `(expr1, expr2, expr3...)`
+#[cfg(not(feature = "parallel"))]
+#[macro_export]
+macro_rules! join {
+    ($($e: expr),+) => {
+        ($($e),+)
+    };
+}
+
+/// `(a, (b, c)) => (a, b, c)`
+#[macro_export]
+macro_rules! unnest_tuple {
+    ($a: expr => $v: expr) => {{
+        $v
+    }};
+    ($a: expr, $b: expr => $v: expr) => {{
+        let (_a, _b) = $v;
+
+        (_a, _b)
+    }};
+    ($a: expr, $b: expr, $c: expr => $v: expr) => {{
+        let (_a, (_b, _c)) = $v;
+
+        (_a, _b, _c)
+    }};
+    ($a: expr, $b: expr, $c: expr, $d: expr => $v: expr) => {{
+        let (_a, (_b, (_c, _d))) = $v;
+
+        (_a, _b, _c, _d)
+    }};
+    ($a: expr, $b: expr, $c: expr, $d: expr, $e: expr => $v: expr) => {{
+        let (_a, (_b, (_c, (_d, _e)))) = $v;
+
+        (_a, _b, _c, _d, _e)
+    }};
+    ($a: expr, $b: expr, $c: expr, $d: expr, $e: expr, $f: expr => $v: expr) => {{
+        let (_a, (_b, (_c, (_d, (_e, _f))))) = $v;
+
+        (_a, _b, _c, _d, _e, _f)
+    }};
+}
+
+/// `impl Iterator` or `impl ParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(feature = "parallel")]
+macro_rules! impl_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl rayon::prelude::ParallelIterator<Item = $item> $($tt)* }
+}
+
+/// `impl Iterator` or `impl ParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(not(feature = "parallel"))]
+macro_rules! impl_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl core::iter::Iterator<Item = $item> $($tt)* }
+}
+
+/// `impl IntoIterator` or `impl IntoParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(feature = "parallel")]
+macro_rules! impl_into_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl rayon::prelude::IntoParallelIterator<Item = $item> $($tt)* }
+}
+
+/// `impl IntoIterator` or `impl IntoParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(not(feature = "parallel"))]
+macro_rules! impl_into_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl core::iter::IntoIterator<Item = $item> $($tt)* }
+}
+
+/// `impl DoubleEndedIterator + ExactSizeIterator` or `impl IndexedParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(feature = "parallel")]
+macro_rules! impl_indexed_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl rayon::prelude::IndexedParallelIterator<Item = $item> $($tt)* }
+}
+
+/// `impl DoubleEndedIterator + ExactSizeIterator` or `impl IndexedParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(not(feature = "parallel"))]
+macro_rules! impl_indexed_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl $crate::helpers::DoubleEndedExactSizeIterator<Item = $item> $($tt)* }
+}
+
+/// `impl IntoIterator` where `IntoIter: DoubleEndedIterator + ExactSizeIterator`  or `impl IntoParallelIterator` where `Iter: IndexedParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(feature = "parallel")]
+macro_rules! impl_into_indexed_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl rayon::prelude::IntoParallelIterator<Item = $item, Iter = impl rayon::prelude::IndexedParallelIterator<Item = $item> $($tt)*> $($tt)* }
+}
+
+/// `impl IntoIterator` where `IntoIter: DoubleEndedIterator + ExactSizeIterator`  or `impl IntoParallelIterator` where `Iter: IndexedParallelIterator` depending on the `parallel` feature.
+#[macro_export]
+#[cfg(not(feature = "parallel"))]
+macro_rules! impl_into_indexed_iter {
+    (<Item = $item: ty> $($tt: tt)*) => { impl core::iter::IntoIterator<Item = $item, IntoIter = impl $crate::helpers::DoubleEndedExactSizeIterator<Item = $item> $($tt)*> $($tt)* }
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn unnest_tuple() {
+        let a = unnest_tuple!(1 => 1);
+        assert_eq!([a], [1]);
+        let (a, b) = unnest_tuple!(_a, _b => (1, 2));
+        assert_eq!([a, b], [1, 2]);
+        let (a, b, c) = unnest_tuple!(_a, _b, _c => (1, (2, 3)));
+        assert_eq!([a, b, c], [1, 2, 3]);
+        let (a, b, c, d) = unnest_tuple!(_a, _b, _c, _d => (1, (2, (3, 4))));
+        assert_eq!([a, b, c, d], [1, 2, 3, 4]);
+        let (a, b, c, d, e) = unnest_tuple!(_a, _b, _c, _d, _e => (1, (2, (3, (4, 5)))));
+        assert_eq!([a, b, c, d, e], [1, 2, 3, 4, 5]);
+        let (a, b, c, d, e, f) =
+            unnest_tuple!(_a, _b, _c, _d, _e, _f => (1, (2, (3, (4, (5, 6))))));
+        assert_eq!([a, b, c, d, e, f], [1, 2, 3, 4, 5, 6]);
+    }
+
+    #[test]
+    fn join() {
+        let a = join!(1);
+        assert_eq!([a], [1]);
+        let (a, b) = join!(1, 2);
+        assert_eq!([a, b], [1, 2]);
+        let (a, b, c) = join!(1, 2, 3);
+        assert_eq!([a, b, c], [1, 2, 3]);
+        let (a, b, c, d) = join!(1, 2, 3, 4);
+        assert_eq!([a, b, c, d], [1, 2, 3, 4]);
+        let (a, b, c, d, e) = join!(1, 2, 3, 4, 5);
+        assert_eq!([a, b, c, d, e], [1, 2, 3, 4, 5]);
+        let (a, b, c, d, e, f) = join!(1, 2, 3, 4, 5, 6);
+        assert_eq!([a, b, c, d, e, f], [1, 2, 3, 4, 5, 6]);
+    }
+}
